+++ main/CMakeLists.txt
@@ -1,6 +1,5 @@
-set(COMPONENT_SRCS 
+set(COMPONENT_SRCS
     "main.cpp"
-    "wifiswitcher.cpp"
 )
 
 set(COMPONENT_ADD_INCLUDEDIRS "")
