--- main/wifiswitcher.hpp
@@ -1,36 +1,0 @@
-#ifndef WIFISWITCHER_HPP_
-#define WIFISWITCHER_HPP_
-
-#include "event_handler.hpp"
-
-#include "wifiap.hpp"
-#include "wifistation.hpp"
-
-#define WIFI_MAX_CONNECTION_RETRY 4
-
-class WifiSwitcher : public EventHandler
-{
-public:
-  
-  static WifiAP* ap();
-  static WifiStation* station();
-
-  static void start();
-
-  esp_err_t handle_event(void *, system_event_t *);
-
-  static WifiSwitcher* instance();
-
-private:
-  WifiSwitcher();
-
-  static WifiSwitcher* instance_;
-
-  WifiAP* ap_;
-  WifiStation* station_;
-
-  uint8_t tries_ = 0;
-};
-
-
-#endif
